Deprecate TimeZone serialization
* Remove the `@kotlinx.serialization.Serializable`
  annotation from `TimeZone` and `FixedOffsetTimeZone`.
* Deprecate `TimeZoneSerializer`
  and `FixedOffsetTimeZoneSerializer`
  with a warning.
* Introduce and deprecate the `.serializer()`
  functions in their companion objects with a warning.
  Previously, this function used to be generated by the
  `kotlinx.serialization` compiler plugin.

Fixes #576
