+++ integration-testing/serialization/common/test/TimeZoneSerializationTest.kt
@@ -34,23 +34,15 @@ class TimeZoneSerializationTest {
         }
     }
 
+    @Suppress("DEPRECATION")
     @Test
     fun testZoneOffsetSerialization() {
         zoneOffsetSerialization(FixedOffsetTimeZoneSerializer)
     }
 
+    @Suppress("DEPRECATION")
     @Test
     fun testSerialization() {
         serialization(TimeZoneSerializer)
-    }
-
-    @Test
-    fun testDefaultSerializers() {
-        assertKSerializerName<FixedOffsetTimeZone>(
-            "kotlinx.datetime.FixedOffsetTimeZone", Json.serializersModule.serializer()
-        )
-        zoneOffsetSerialization(Json.serializersModule.serializer())
-        assertKSerializerName<TimeZone>("kotlinx.datetime.TimeZone", Json.serializersModule.serializer())
-        serialization(Json.serializersModule.serializer())
     }
 }
